% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap4R.R
\docType{package}
\name{ldap4R}
\alias{ldap4R}
\alias{ldap4R-package}
\title{R interface to LDAP}
\description{
R interface to LDAP
}
\details{
\tabular{ll}{
   Package: \tab ldap4R\cr
   Type: \tab Package\cr
   Version
   : \tab 0.0.9000\cr
   Date: \tab 2019-10-29\cr
   License: \tab MIT\cr
   LazyLoad: \tab yes\cr
 }
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
