% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAPClient.R
\docType{class}
\name{LDAPClient}
\alias{LDAPClient}
\title{LDAPClient}
\format{\code{\link{R6Class}} object.}
\usage{
LDAPClient
}
\value{
Object of \code{\link{R6Class}} with methods for communication a LDAP.
}
\description{
LDAPClient
}
\section{Fields}{

\describe{
\item{\code{url}}{URL of the LDAP}

\item{\code{root}}{root LDAP DN}
}}

\section{Methods}{

\describe{
 \item{\code{new(hostname, port, dc, logger)}}{
   This method is used to instantiate a LDAP client.  The logger can be either
   NULL, "INFO" (with minimum logs), or "DEBUG" (for complete curl calls logs)
 }
 \item{\code{getPersons(pretty)}}{
   Retrieves a list of persons. By default, this is returned as \code{data.frame}.
   Set \code{pretty = FALSE} to return a list of \link{LDAPPerson} objects.
 }
}
}

\examples{
\dontrun{
  ldap <- LDAPClient$new(hostname = "localhost", port = 389, dc = "planetexpress.com")
}

}
\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ldap}
