% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAPObject.R
\docType{class}
\name{LDAPObject}
\alias{LDAPObject}
\title{LDAPObject}
\format{\code{\link{R6Class}} object.}
\usage{
LDAPObject
}
\value{
Object of \code{\link{R6Class}} representing a LDAP object.
}
\description{
LDAPObject
}
\note{
abstract class
}
\section{Methods}{

\describe{
 \item{\code{new(dn)}}{
   This method is used to instantiate a LDAP object from its distinguished name (DN).
 }
 \item{\code{parseDN(dn)}}{
   Parses a distinguished name into \code{list}
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ldap}
\keyword{object}
