% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAPPerson.R
\docType{class}
\name{LDAPPerson}
\alias{LDAPPerson}
\title{LDAPPerson}
\format{\code{\link{R6Class}} object.}
\usage{
LDAPPerson
}
\value{
Object of \code{\link{R6Class}} representing a LDAP person.
}
\description{
LDAPPerson
}
\section{Fields}{

\describe{
\item{\code{objectClass}}{}
}}

\section{Methods}{

\describe{
 \item{\code{new(dn)}}{
   This method is used to instantiate a LDAP person from its distinguished name (DN).
 }
}
}

\author{
Emmanuel Blondel <emmanuel.blondel1@gmail.com>
}
\keyword{ldap}
